refactor(hub-discussions): tests for ChannelPermissions class
affects: @esri/hub-discussions
