+++ packages/discussions/test/utils/channels/can-create-channel.test.ts
@@ -7,6 +7,7 @@ import {
   Role,
   SharingAccess,
 } from "../../../src/types";
+import { ChannelPermission } from "../../../src/utils/channel-permission";
 import { canCreateChannel } from "../../../src/utils/channels";
 import { CANNOT_DISCUSS } from "../../../src/utils/constants";
 
@@ -68,279 +69,48 @@ function buildCompleteAcl() {
 }
 
 describe("canCreateChannel", () => {
-  describe("With ChannelAcl Permissions", () => {
-    it("returns false if user not authenticated", () => {
+  describe("with channelAcl", () => {
+    let canCreateChannelSpy: jasmine.Spy;
+
+    beforeAll(() => {
+      canCreateChannelSpy = spyOn(
+        ChannelPermission.prototype,
+        "canCreateChannel"
+      );
+    });
+
+    beforeEach(() => {
+      canCreateChannelSpy.calls.reset();
+    });
+
+    it("return true if channelPermission.canCreateChannel is true", () => {
+      canCreateChannelSpy.and.callFake(() => true);
+
+      const user = buildUser();
       const channel = {
         channelAcl: [{ category: AclCategory.ANONYMOUS_USER, role: Role.READ }],
       } as IChannel;
-      const user = buildUser({ username: null });
 
-      expect(canCreateChannel(channel, user)).toEqual(false);
+      expect(canCreateChannel(channel, user)).toBe(true);
+
+      expect(canCreateChannelSpy.calls.count()).toBe(1);
+      const [arg] = canCreateChannelSpy.calls.allArgs()[0]; // arg for 1st call
+      expect(arg).toBe(user);
     });
 
-    it("returns false if permissions list is empty", () => {
-      const channel = { channelAcl: [] } as unknown as IChannel;
+    it("return false if channelPermission.canCreateChannel is false", () => {
+      canCreateChannelSpy.and.callFake(() => false);
+
       const user = buildUser();
+      const channel = {
+        channelAcl: [{ category: AclCategory.ANONYMOUS_USER, role: Role.READ }],
+      } as IChannel;
 
-      expect(canCreateChannel(channel, user)).toEqual(false);
-    });
+      expect(canCreateChannel(channel, user)).toBe(false);
 
-    it("returns true with a full valid channelAcl if user is the org_admin, in all groups, and org permissions only relate to users org", () => {
-      const channel = { channelAcl: buildCompleteAcl() } as IChannel;
-      const user = buildUser({ role: "org_admin" });
-
-      expect(canCreateChannel(channel, user)).toEqual(true);
-    });
-
-    describe("Anonymous User Permissions", () => {
-      it("returns true if user is org_admin", () => {
-        const channel = {
-          channelAcl: [
-            { category: AclCategory.ANONYMOUS_USER, role: Role.READ },
-          ],
-        } as IChannel;
-        const user = buildUser({ role: "org_admin" });
-
-        expect(canCreateChannel(channel, user)).toEqual(true);
-      });
-
-      it("returns false if user is not org_admin", () => {
-        const channel = {
-          channelAcl: [
-            { category: AclCategory.ANONYMOUS_USER, role: Role.READ },
-          ],
-        } as IChannel;
-        const user = buildUser();
-
-        expect(canCreateChannel(channel, user)).toEqual(false);
-      });
-    });
-
-    describe("Authenticated User Permissions", () => {
-      it("returns true if user is org_admin", () => {
-        const channel = {
-          channelAcl: [
-            { category: AclCategory.AUTHENTICATED_USER, role: Role.READ },
-          ],
-        } as IChannel;
-        const user = buildUser({ role: "org_admin" });
-
-        expect(canCreateChannel(channel, user)).toEqual(true);
-      });
-
-      it("returns false if user is not org_admin", () => {
-        const channel = {
-          channelAcl: [
-            { category: AclCategory.AUTHENTICATED_USER, role: Role.READ },
-          ],
-        } as IChannel;
-        const user = buildUser();
-
-        expect(canCreateChannel(channel, user)).toEqual(false);
-      });
-    });
-
-    describe("Group Permissions", () => {
-      it("returns true if user is a member of all groups and memberType is member, admin, or owner", async () => {
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId1,
-              role: Role.OWNER,
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.MEMBER,
-              key: groupId2,
-              role: Role.READ,
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.MEMBER,
-              key: groupId3,
-              role: Role.READ,
-            },
-          ],
-        } as IChannel;
-        const user = buildUser({
-          groups: [
-            buildGroup(groupId1, "member"),
-            buildGroup(groupId2, "admin"),
-            buildGroup(groupId3, "owner"),
-          ],
-        });
-
-        expect(canCreateChannel(channel, user)).toEqual(true);
-      });
-
-      it("returns false is user is a member of all groups and memberType is none", async () => {
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId1,
-              role: Role.OWNER,
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.MEMBER,
-              key: groupId2,
-              role: Role.READ,
-            },
-          ],
-        } as IChannel;
-        const user = buildUser({
-          groups: [buildGroup(groupId1, "admin"), buildGroup(groupId2, "none")],
-        });
-
-        expect(canCreateChannel(channel, user)).toEqual(false);
-      });
-
-      it("returns false is user is a member of all groups but group is not discussable", async () => {
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId1,
-              role: Role.OWNER,
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.MEMBER,
-              key: groupId2,
-              role: Role.READ,
-            },
-          ],
-        } as IChannel;
-        const user = buildUser({
-          groups: [
-            buildGroup(groupId1, "admin"),
-            buildGroup(groupId2, "member", [CANNOT_DISCUSS]),
-          ],
-        });
-
-        expect(canCreateChannel(channel, user)).toEqual(false);
-      });
-
-      it("returns false if user is not a member of every group", async () => {
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId1,
-              role: Role.READWRITE,
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId2,
-              role: Role.READWRITE,
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId3, // user not member here
-              role: Role.READWRITE,
-            },
-          ],
-        } as IChannel;
-        const user = buildUser();
-
-        expect(canCreateChannel(channel, user)).toEqual(false);
-      });
-    });
-
-    describe("Org Permissions", () => {
-      it("returns true if user is org_admin and every permission is for the users orgId", () => {
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.ORG,
-              subCategory: AclSubCategory.ADMIN,
-              key: orgId1,
-              role: Role.OWNER,
-            },
-            {
-              category: AclCategory.ORG,
-              subCategory: AclSubCategory.MEMBER,
-              key: orgId1,
-              role: Role.READ,
-            },
-          ],
-        } as IChannel;
-        const user = buildUser({ role: "org_admin" });
-
-        expect(canCreateChannel(channel, user)).toEqual(true);
-      });
-
-      it("returns false if user is org_admin and every permission is not for the users orgId", () => {
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.ORG,
-              subCategory: AclSubCategory.ADMIN,
-              key: orgId1,
-              role: Role.OWNER,
-            },
-            {
-              category: AclCategory.ORG,
-              subCategory: AclSubCategory.MEMBER,
-              key: "unknown",
-              role: Role.READ,
-            },
-          ],
-        } as IChannel;
-        const user = buildUser({ role: "org_admin" });
-
-        expect(canCreateChannel(channel, user)).toEqual(false);
-      });
-
-      it("returns false if user is not org_admin", () => {
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.ORG,
-              subCategory: AclSubCategory.ADMIN,
-              key: orgId1,
-              role: Role.OWNER,
-            },
-            {
-              category: AclCategory.ORG,
-              subCategory: AclSubCategory.MEMBER,
-              key: orgId1,
-              role: Role.READ,
-            },
-          ],
-        } as IChannel;
-        const user = buildUser();
-
-        expect(canCreateChannel(channel, user)).toEqual(false);
-      });
-    });
-
-    describe("User Permissions", () => {
-      it("returns false if user permissions are included", () => {
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.USER,
-              key: "johnUsername",
-              role: Role.READ,
-            },
-          ],
-        } as IChannel;
-        const user = buildUser({ role: "org_admin" });
-
-        expect(canCreateChannel(channel, user)).toEqual(false);
-      });
+      expect(canCreateChannelSpy.calls.count()).toBe(1);
     });
   });
-
   describe("With Legacy Permissions", () => {
     it("returns false if user not authenticated", () => {
       const channel = {
