+++ packages/discussions/test/utils/channels/can-post-to-channel.test.ts
@@ -1,23 +1,12 @@
 import { IGroup } from "@esri/arcgis-rest-types";
 import {
   AclCategory,
-  AclSubCategory,
   IChannel,
   IDiscussionsUser,
   Role,
 } from "../../../src/types";
+import { ChannelPermission } from "../../../src/utils/channel-permission";
 import { canPostToChannel } from "../../../src/utils/channels/can-post-to-channel";
-import { CANNOT_DISCUSS } from "../../../src/utils/constants";
-
-const ALLOWED_GROUP_ROLES = Object.freeze(["owner", "admin", "member"]);
-
-const ALLOWED_ROLES_FOR_POSTING = Object.freeze([
-  Role.WRITE,
-  Role.READWRITE,
-  Role.MANAGE,
-  Role.MODERATE,
-  Role.OWNER,
-]);
 
 const orgId1 = "3ef";
 const groupId1 = "aaa";
@@ -463,416 +452,45 @@ describe("canPostToChannel", () => {
   });
 
   describe("with channelAcl", () => {
-    it("returns false if user logged in and channel permissions are empty", async () => {
+    let canPostToChannelSpy: jasmine.Spy;
+
+    beforeAll(() => {
+      canPostToChannelSpy = spyOn(
+        ChannelPermission.prototype,
+        "canPostToChannel"
+      );
+    });
+
+    beforeEach(() => {
+      canPostToChannelSpy.calls.reset();
+    });
+
+    it("return true if channelPermission.canPostToChannel is true", () => {
+      canPostToChannelSpy.and.callFake(() => true);
+
       const user = buildUser();
       const channel = {
-        channelAcl: [],
-      } as any as IChannel;
+        channelAcl: [{ category: AclCategory.ANONYMOUS_USER, role: Role.READ }],
+      } as IChannel;
 
-      expect(canPostToChannel(channel, user)).toBe(false);
+      expect(canPostToChannel(channel, user)).toBe(true);
+
+      expect(canPostToChannelSpy.calls.count()).toBe(1);
+      const [arg] = canPostToChannelSpy.calls.allArgs()[0]; // arg for 1st call
+      expect(arg).toBe(user);
     });
 
-    it("returns false if user not logged in and channel permissions are empty", async () => {
-      const user = buildUser({ username: null });
+    it("return false if channelPermission.canPostToChannel is false", () => {
+      canPostToChannelSpy.and.callFake(() => false);
+
+      const user = buildUser();
       const channel = {
-        channelAcl: [],
-      } as any as IChannel;
+        channelAcl: [{ category: AclCategory.ANONYMOUS_USER, role: Role.READ }],
+      } as IChannel;
 
       expect(canPostToChannel(channel, user)).toBe(false);
-    });
 
-    describe("any anonymous user channel permissions", () => {
-      it(`returns true if anonymous permission defined and role is allowed`, () => {
-        const user = buildUser({ username: null });
-
-        ALLOWED_ROLES_FOR_POSTING.forEach((allowedRole) => {
-          const channel = {
-            channelAcl: [
-              { category: AclCategory.ANONYMOUS_USER, role: allowedRole },
-            ],
-          } as IChannel;
-
-          expect(canPostToChannel(channel, user)).toBe(true);
-        });
-      });
-
-      it("returns false if anonymous permission defined but role is read", () => {
-        const user = buildUser({ username: null });
-        const channel = {
-          channelAcl: [
-            { category: AclCategory.ANONYMOUS_USER, role: Role.READ },
-          ],
-        } as IChannel;
-
-        expect(canPostToChannel(channel, user)).toBe(false);
-      });
-    });
-
-    describe("any authenticated user channel permissions", () => {
-      it(`returns true if authenticated permission defined, user logged in, and role is allowed`, async () => {
-        const user = buildUser();
-
-        ALLOWED_ROLES_FOR_POSTING.forEach((allowedRole) => {
-          const channel = {
-            channelAcl: [
-              { category: AclCategory.AUTHENTICATED_USER, role: allowedRole },
-            ],
-          } as IChannel;
-
-          expect(canPostToChannel(channel, user)).toBe(true);
-        });
-      });
-
-      it("returns false if authenticated permission defined, user logged in, and role is read", async () => {
-        const user = buildUser();
-        const channel = {
-          channelAcl: [
-            { category: AclCategory.AUTHENTICATED_USER, role: Role.READ },
-          ],
-        } as IChannel;
-
-        expect(canPostToChannel(channel, user)).toBe(false);
-      });
-
-      it("returns false if authenticated permission defined and user is not logged in", async () => {
-        const user = buildUser({ username: null });
-        const channel = {
-          channelAcl: [
-            { category: AclCategory.AUTHENTICATED_USER, role: Role.READWRITE },
-          ],
-        } as IChannel;
-
-        expect(canPostToChannel(channel, user)).toBe(false);
-      });
-    });
-
-    describe("specific authenticated user channel permissions", () => {
-      it("returns true if user is in permissions list and role is allowed", () => {
-        const user = buildUser();
-
-        ALLOWED_ROLES_FOR_POSTING.forEach((allowedRole) => {
-          const channel = {
-            channelAcl: [
-              {
-                category: AclCategory.USER,
-                key: user.username,
-                role: allowedRole,
-              },
-            ],
-          } as IChannel;
-
-          expect(canPostToChannel(channel, user)).toBe(true);
-        });
-      });
-
-      it("returns false if user is in permissions list but role is read", () => {
-        const user = buildUser();
-        const channel = {
-          channelAcl: [
-            { category: AclCategory.USER, key: user.username, role: Role.READ },
-          ],
-        } as IChannel;
-
-        expect(canPostToChannel(channel, user)).toBe(false);
-      });
-    });
-
-    describe("group channel permissions", () => {
-      it("returns true if user is group member in group permission list and role is allowed", async () => {
-        ALLOWED_ROLES_FOR_POSTING.forEach((allowedRole) => {
-          const channel = {
-            channelAcl: [
-              {
-                category: AclCategory.GROUP,
-                subCategory: AclSubCategory.MEMBER,
-                key: groupId1,
-                role: allowedRole, // members write
-              },
-              {
-                category: AclCategory.GROUP,
-                subCategory: AclSubCategory.ADMIN,
-                key: groupId1,
-                role: Role.READ,
-              },
-            ],
-          } as IChannel;
-
-          ALLOWED_GROUP_ROLES.forEach((memberType) => {
-            const user = buildUser({
-              orgId: orgId1,
-              groups: [buildGroup(groupId1, memberType)], // member in groupId1
-            });
-
-            expect(canPostToChannel(channel, user)).toBe(true);
-          });
-        });
-      });
-
-      it("returns false if user is group member in group permission list and role is NOT allowed", async () => {
-        const user = buildUser(); // member in groupId1
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.MEMBER,
-              key: groupId1,
-              role: Role.READ, // members read
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId1,
-              role: Role.READ,
-            },
-          ],
-        } as IChannel;
-
-        expect(canPostToChannel(channel, user)).toBe(false);
-      });
-
-      it("returns false if user is group member in group permission list, role is allowed, but userMemberType is none", async () => {
-        const user = buildUser({
-          groups: [buildGroup(groupId1, "none")], // none in groupId1
-        });
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.MEMBER,
-              key: groupId1,
-              role: Role.READWRITE, // members read
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId1,
-              role: Role.READWRITE, // admins read
-            },
-          ],
-        } as IChannel;
-
-        expect(canPostToChannel(channel, user)).toBe(false);
-      });
-
-      it("returns true if user is group owner/admin in group permission list and role is allowed", async () => {
-        ALLOWED_ROLES_FOR_POSTING.forEach((allowedRole) => {
-          const channel = {
-            channelAcl: [
-              {
-                category: AclCategory.GROUP,
-                subCategory: AclSubCategory.MEMBER,
-                key: groupId1,
-                role: Role.READ,
-              },
-              {
-                category: AclCategory.GROUP,
-                subCategory: AclSubCategory.ADMIN,
-                key: groupId1,
-                role: allowedRole, // admins write
-              },
-            ],
-          } as IChannel;
-
-          ["owner", "admin"].forEach((memberType) => {
-            const user = buildUser({
-              orgId: orgId1,
-              groups: [buildGroup(groupId1, memberType)], // admin in groupId1
-            });
-
-            expect(canPostToChannel(channel, user)).toBe(true);
-          });
-        });
-      });
-
-      it("returns false if user is group owner/admin in group permission list and role is NOT allowed", async () => {
-        const user = buildUser(); // admin in groupId2
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.MEMBER,
-              key: groupId2,
-              role: Role.READ,
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId2,
-              role: Role.READ, // admins read
-            },
-          ],
-        } as IChannel;
-
-        expect(canPostToChannel(channel, user)).toBe(false);
-      });
-
-      it("returns true if user is group member of at least one group in permissions list that is discussable", async () => {
-        const user = buildUser({
-          orgId: orgId1,
-          groups: [
-            buildGroup(groupId1, "member"), // member in groupId1
-            buildGroup(groupId2, "member", [CANNOT_DISCUSS]), // member in groupId2
-          ],
-        });
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.MEMBER,
-              key: groupId1,
-              role: Role.READWRITE, // members write
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId1,
-              role: Role.READ,
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.MEMBER,
-              key: groupId2,
-              role: Role.READWRITE, // members write, group CANNOT_DISCUSS
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId2,
-              role: Role.READ,
-            },
-          ],
-        } as IChannel;
-
-        expect(canPostToChannel(channel, user)).toBe(true);
-      });
-
-      it("returns false if user is group member in permissions list but the group is not discussable", async () => {
-        const user = buildUser({
-          orgId: orgId1,
-          groups: [
-            buildGroup(groupId1, "member", [CANNOT_DISCUSS]), // member in groupId1
-          ],
-        });
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.MEMBER,
-              key: groupId1,
-              role: Role.READWRITE, // members write
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId1,
-              role: Role.READ,
-            },
-          ],
-        } as IChannel;
-
-        expect(canPostToChannel(channel, user)).toBe(false);
-      });
-
-      it("returns false if user is group admin but group is not in permissions list", async () => {
-        const user = buildUser({
-          orgId: orgId1,
-          groups: [
-            buildGroup("unknownGroupId", "admin"), // member in unknownGroupId
-          ],
-        });
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.MEMBER,
-              key: groupId1,
-              role: Role.READWRITE, // members write
-            },
-            {
-              category: AclCategory.GROUP,
-              subCategory: AclSubCategory.ADMIN,
-              key: groupId1,
-              role: Role.READWRITE, // admin write
-            },
-          ],
-        } as IChannel;
-
-        expect(canPostToChannel(channel, user)).toBe(false);
-      });
-    });
-
-    describe("org channel permissions", () => {
-      it("returns true if user is org member in permissions list and role is allowed", async () => {
-        const user = buildUser();
-
-        ALLOWED_ROLES_FOR_POSTING.forEach((allowedRole) => {
-          const channel = {
-            channelAcl: [
-              {
-                category: AclCategory.ORG,
-                subCategory: AclSubCategory.MEMBER,
-                key: user.orgId,
-                role: allowedRole, // members write
-              },
-              {
-                category: AclCategory.ORG,
-                subCategory: AclSubCategory.ADMIN,
-                key: user.orgId,
-                role: Role.READ, // admin read
-              },
-            ],
-          } as IChannel;
-
-          expect(canPostToChannel(channel, user)).toBe(true);
-        });
-      });
-
-      it("returns true if user is org_admin in permissions list and role is allowed", async () => {
-        const user = buildUser({ role: "org_admin" });
-
-        ALLOWED_ROLES_FOR_POSTING.forEach((allowedRole) => {
-          const channel = {
-            channelAcl: [
-              {
-                category: AclCategory.ORG,
-                subCategory: AclSubCategory.MEMBER,
-                key: user.orgId,
-                role: Role.READ, // members read
-              },
-              {
-                category: AclCategory.ORG,
-                subCategory: AclSubCategory.ADMIN,
-                key: user.orgId,
-                role: Role.READWRITE, // admin write
-              },
-            ],
-          } as IChannel;
-
-          expect(canPostToChannel(channel, user)).toBe(true);
-        });
-      });
-
-      it("returns false if user is not in the org", async () => {
-        const user = buildUser({ orgId: "unknownOrgId" });
-        const channel = {
-          channelAcl: [
-            {
-              category: AclCategory.ORG,
-              subCategory: AclSubCategory.MEMBER,
-              key: orgId1,
-              role: Role.READ, // members read
-            },
-            {
-              category: AclCategory.ORG,
-              subCategory: AclSubCategory.ADMIN,
-              key: orgId1,
-              role: Role.READWRITE, // admin write
-            },
-          ],
-        } as IChannel;
-
-        expect(canPostToChannel(channel, user)).toBe(false);
-      });
+      expect(canPostToChannelSpy.calls.count()).toBe(1);
     });
   });
